requirements.txt
openai-whisper
ffmpeg-python
deep-translator